move file up one level
