+++ rubin_sim/maf/mafContrib/depthLimitedNumGalMetric.py
