--- rubin_sim/maf/mafContrib/lssmetrics/__init__.py
@@ -1,1 +1,0 @@
-from .depthLimitedNumGalMetric import *
